requirements.txt
protobuf==3.0.0
ctypes_snappy==1.03
